requirements.txt
beautifulsoup4==4.7.1
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.2
cycler==0.10.0
demjson==2.2.4
draftfast==2.4.2
entrypoints==0.3
flake8==3.7.5
google-apputils==0.4.2
idna==2.6
kiwisolver==1.0.1
matplotlib==3.0.2
mccabe==0.6.1
mysql-connector-python-rf==2.2.2
nose==1.3.7
numpy==1.16.1
ortools==6.5.4527
protobuf==3.3.0
PuLP==1.6.8
pycodestyle==2.5.0
pydfs-lineup-optimizer==1.3.1
pyflakes==2.1.0
pyparsing==2.2.0
python-dateutil==2.8.0
python-gflags==3.1.2
pytz==2018.9
requests==2.18.4
scikit-learn==0.19.1
scipy==1.2.0
six==1.12.0
soupsieve==1.7.3
terminaltables==3.1.0
testpath==0.4.2
urllib3==1.22
